requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
pyjwt
streamlit
requests
python-multipart
watchdog
passlib